refactor: Make the InfectionVersion more usable
