+++ tests/phpunit/AutoReview/ProjectCode/ProjectCodeProvider.php
@@ -57,7 +57,6 @@ use Infection\Configuration\Schema\SchemaValidator;
 use Infection\Configuration\SourceFilter\FakeSourceFilter;
 use Infection\Configuration\SourceFilter\GitDiffFilter;
 use Infection\Configuration\SourceFilter\IncompleteGitDiffFilter;
-use Infection\Console\Application;
 use Infection\Console\XdebugHandler;
 use Infection\Differ\Tokens;
 use Infection\Event\Events\MutationAnalysis\MutationEvaluation\MutationEvaluationWasStarted;
@@ -133,7 +132,6 @@ final class ProjectCodeProvider
      */
     public const array NON_TESTED_CONCRETE_CLASSES = [
         AdapterInstaller::class,
-        Application::class,
         BaseMutatorTestCase::class,
         BaseOption::class,
         ConcreteComposerExecutableFinder::class,
